+++ baseclasses/utils/__init__.py
@@ -1,5 +1,5 @@
 from .containers import CaseInsensitiveSet, CaseInsensitiveDict
 from .error import Error
-from .py3Util import getPy3SafeString
+from .utils import getPy3SafeString
 
 __all__ = ["CaseInsensitiveSet", "CaseInsensitiveDict", "Error", "getPy3SafeString"]
